+++ tomcat/src/main/java/com/techcourse/controller/StaticFileController.java
@@ -1,13 +1,13 @@
 package com.techcourse.controller;
 
+import com.techcourse.utility.StaticFileUtility;
 import java.util.List;
-import org.apache.catalina.container.controller.AbstractController;
-import org.apache.catalina.container.http.request.HttpRequest;
-import org.apache.catalina.container.http.response.HttpResponse;
-import org.apache.catalina.container.http.value.ContentType;
-import org.apache.catalina.container.http.value.HttpHeader;
-import org.apache.catalina.container.http.value.StatusCode;
-import org.apache.catalina.container.reader.StaticFileUtility;
+import org.apache.catalina.core.controller.AbstractController;
+import org.apache.tomcat.util.http.request.HttpRequest;
+import org.apache.tomcat.util.http.response.HttpResponse;
+import org.apache.tomcat.util.http.value.ContentType;
+import org.apache.tomcat.util.http.value.HttpHeader;
+import org.apache.tomcat.util.http.value.StatusCode;
 
 public class StaticFileController extends AbstractController {
