--- tomcat/src/main/java/org/apache/catalina/container/controller/AbstractController.java
@@ -1,34 +1,0 @@
-package org.apache.catalina.container.controller;
-
-import org.apache.catalina.container.http.request.HttpRequest;
-import org.apache.catalina.container.http.response.HttpResponse;
-import org.apache.catalina.container.http.value.HttpMethod;
-
-public abstract class AbstractController implements Controller {
-
-    private final String targetUrl;
-
-    public AbstractController(String targetUrl) {
-        this.targetUrl = targetUrl;
-    }
-
-    @Override
-    public boolean canProcessable(HttpRequest request) {
-        return request.getUri().equals(targetUrl);
-    }
-
-    @Override
-    public void service(HttpRequest request, HttpResponse response) throws Exception {
-        if (request.getMethod() == HttpMethod.GET) {
-            doGet(request, response);
-        } else if (request.getMethod() == HttpMethod.POST) {
-            doPost(request, response);
-        }
-    }
-
-    protected void doPost(HttpRequest request, HttpResponse response) throws Exception {
-    }
-
-    protected void doGet(HttpRequest request, HttpResponse response) throws Exception {
-    }
-}
