--- tomcat/src/main/java/org/apache/catalina/container/controller/Controller.java
@@ -1,11 +1,0 @@
-package org.apache.catalina.container.controller;
-
-import org.apache.catalina.container.http.request.HttpRequest;
-import org.apache.catalina.container.http.response.HttpResponse;
-
-public interface Controller {
-
-    boolean canProcessable(HttpRequest request);
-
-    void service(HttpRequest request, HttpResponse response) throws Exception;
-}
