--- tomcat/src/main/java/org/apache/catalina/container/exception/DataNotFoundException.java
@@ -1,8 +1,0 @@
-package org.apache.catalina.container.exception;
-
-public class DataNotFoundException extends RuntimeException {
-
-    public DataNotFoundException(String message) {
-        super(message);
-    }
-}
