--- tomcat/src/main/java/org/apache/catalina/container/exception/FileReadException.java
@@ -1,8 +1,0 @@
-package org.apache.catalina.container.exception;
-
-public class FileReadException extends RuntimeException {
-
-    public FileReadException(String message) {
-        super(message);
-    }
-}
