--- tomcat/src/main/java/org/apache/catalina/container/exception/InvalidRequestException.java
@@ -1,8 +1,0 @@
-package org.apache.catalina.container.exception;
-
-public class InvalidRequestException extends RuntimeException {
-
-    public InvalidRequestException(String message) {
-        super(message);
-    }
-}
