--- tomcat/src/main/java/org/apache/catalina/container/exception/RequestProcessingException.java
@@ -1,8 +1,0 @@
-package org.apache.catalina.container.exception;
-
-public class RequestProcessingException extends RuntimeException {
-
-    public RequestProcessingException(String message) {
-        super(message);
-    }
-}
