--- tomcat/src/main/java/org/apache/catalina/container/exception/SocketReadException.java
@@ -1,8 +1,0 @@
-package org.apache.catalina.container.exception;
-
-public class SocketReadException extends RuntimeException {
-
-    public SocketReadException(String message) {
-        super(message);
-    }
-}
