--- tomcat/src/main/java/org/apache/catalina/container/exception/SocketWriteException.java
@@ -1,8 +1,0 @@
-package org.apache.catalina.container.exception;
-
-public class SocketWriteException extends RuntimeException {
-
-    public SocketWriteException(String message) {
-        super(message);
-    }
-}
