--- tomcat/src/main/java/org/apache/catalina/container/exception/UnauthorizedException.java
@@ -1,8 +1,0 @@
-package org.apache.catalina.container.exception;
-
-public class UnauthorizedException extends RuntimeException {
-
-    public UnauthorizedException(String message) {
-        super(message);
-    }
-}
