--- tomcat/src/main/java/org/apache/catalina/container/http/value/HttpMethod.java
@@ -1,5 +1,0 @@
-package org.apache.catalina.container.http.value;
-
-public enum HttpMethod {
-    GET, POST, PUT, DELETE, PATCH, HEAD, OPTIONS, CONNECT
-}
