--- tomcat/src/main/java/org/apache/catalina/container/http/value/StatusCode.java
@@ -1,26 +1,0 @@
-package org.apache.catalina.container.http.value;
-
-public enum StatusCode {
-
-    OK(200, "OK"),
-    FOUND(302, "FOUND"),
-    BAD_REQUEST(400, "BAD_REQUEST"),
-    NOT_FOUND(404, "NOT_FOUND"),
-    INTERNAL_SERVER_ERROR(500, "INTERNAL_SERVER_ERROR");
-
-    private final int code;
-    private final String message;
-
-    StatusCode(int code, String message) {
-        this.code = code;
-        this.message = message;
-    }
-
-    public int getCode() {
-        return code;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
