+++ tomcat/src/main/java/org/apache/catalina/core/exception/DataNotFoundException.java
@@ -1,0 +1,8 @@
+package org.apache.catalina.core.exception;
+
+public class DataNotFoundException extends RuntimeException {
+
+    public DataNotFoundException(String message) {
+        super(message);
+    }
+}
