+++ tomcat/src/main/java/org/apache/catalina/core/exception/FileReadException.java
@@ -1,0 +1,8 @@
+package org.apache.catalina.core.exception;
+
+public class FileReadException extends RuntimeException {
+
+    public FileReadException(String message) {
+        super(message);
+    }
+}
