+++ tomcat/src/main/java/org/apache/catalina/core/exception/InvalidRequestException.java
@@ -1,0 +1,8 @@
+package org.apache.catalina.core.exception;
+
+public class InvalidRequestException extends RuntimeException {
+
+    public InvalidRequestException(String message) {
+        super(message);
+    }
+}
