+++ tomcat/src/main/java/org/apache/catalina/core/exception/RequestProcessingException.java
@@ -1,0 +1,8 @@
+package org.apache.catalina.core.exception;
+
+public class RequestProcessingException extends RuntimeException {
+
+    public RequestProcessingException(String message) {
+        super(message);
+    }
+}
