+++ tomcat/src/main/java/org/apache/catalina/core/exception/SocketReadException.java
@@ -1,0 +1,8 @@
+package org.apache.catalina.core.exception;
+
+public class SocketReadException extends RuntimeException {
+
+    public SocketReadException(String message) {
+        super(message);
+    }
+}
