+++ tomcat/src/main/java/org/apache/catalina/core/exception/SocketWriteException.java
@@ -1,0 +1,8 @@
+package org.apache.catalina.core.exception;
+
+public class SocketWriteException extends RuntimeException {
+
+    public SocketWriteException(String message) {
+        super(message);
+    }
+}
