+++ tomcat/src/main/java/org/apache/catalina/core/exception/UnauthorizedException.java
@@ -1,0 +1,8 @@
+package org.apache.catalina.core.exception;
+
+public class UnauthorizedException extends RuntimeException {
+
+    public UnauthorizedException(String message) {
+        super(message);
+    }
+}
